requirements.txt
pypdf
chromadb
langchain
langchain-openai
langchain-community
langchain-chroma
langchain-text-splitters
python-dotenv
streamlit
tiktoken
